# openapi.yaml
openapi: 3.1.0
info:
    title: Inventory Control API
    description: Per-warehouse inventory management with stock transfers, reservations, purchase orders, and audit logging.
    version: 2.0.0
    license:
        name: MIT
servers:
    - url: http://localhost:3333
      description: Local development
security:
    - bearerAuth: []
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        Error:
            type: object
            properties:
                error:
                    type: string
            required: [error]
        Pagination:
            type: object
            properties:
                page:
                    type: integer
                limit:
                    type: integer
                total:
                    type: integer
                totalPages:
                    type: integer
...
        InventoryItem:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                warehouse_id:
                    type: string
                    format: uuid
                product_id:
                    type: string
                    format: uuid
                quantity:
                    type: integer
                reserved_quantity:
                    type: integer
                updated_at:
                    type: string
                    format: date-time
                warehouse_code:
                    type: string
                warehouse_name:
                    type: string
                sku:
                    type: string
                product_name:
                    type: string
                unit:
                    type: string
                available_stock:
                    type: integer
...
    parameters:
        PageParam:
            name: page
            in: query
            schema:
                type: integer
                default: 1
                minimum: 1
        LimitParam:
            name: limit
            in: query
            schema:
                type: integer
                default: 20
                minimum: 1
                maximum: 100
paths:
...
    /api/v1/inventory/low-stock:
        get:
            summary: Low stock items
            operationId: getLowStock
            tags: [Inventory]
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/LimitParam'
            responses:
                '200':
                    description: Paginated low stock list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/InventoryItem'
                                    pagination:
                                        $ref: '#/components/schemas/Pagination'
...
    /api/v1/inventory/receive:
        post:
            summary: Receive stock
            operationId: receiveStock
            tags: [Inventory]
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [warehouseId, productId, quantity]
                            properties:
                                warehouseId:
                                    type: string
                                    format: uuid
                                productId:
                                    type: string
                                    format: uuid
                                quantity:
                                    type: integer
                                    minimum: 1
                                referenceType:
                                    type: string
                                referenceId:
                                    type: string
                                    format: uuid
                                reason:
                                    type: string
            responses:
                '201':
                    description: Stock received
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: Stock received
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/v1/inventory/issue:
        post:
            summary: Issue stock
            operationId: issueStock
            tags: [Inventory]
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [warehouseId, productId, quantity]
                            properties:
                                warehouseId:
                                    type: string
                                    format: uuid
                                productId:
                                    type: string
                                    format: uuid
                                quantity:
                                    type: integer
                                    minimum: 1
                                reason:
                                    type: string
            responses:
                '201':
                    description: Stock issued
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: Stock issued
                '400':
                    description: Validation error or insufficient stock
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/v1/inventory/adjust:
        post:
            summary: Adjust stock
            operationId: adjustStock
            tags: [Inventory]
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [warehouseId, productId, quantity, type]
                            properties:
                                warehouseId:
                                    type: string
                                    format: uuid
                                productId:
                                    type: string
                                    format: uuid
                                quantity:
                                    type: integer
                                    minimum: 1
                                type:
                                    type: string
                                    enum: [ADJUSTMENT_IN, ADJUSTMENT_OUT]
                                reason:
                                    type: string
            responses:
                '201':
                    description: Stock adjusted
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: Stock adjusted
                '400':
                    description: Validation error or insufficient stock
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...